added update operation

+++ src/main/java/com/blog/repository/UserRepository.java
@@ -8,6 +8,7 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.UUID;
 
 @Slf4j
 @Repository
@@ -39,11 +40,11 @@ public class UserRepository implements CrudRepository<User> {
     }
 
     @Override
-    public ResultSet getById(int id) {
+    public ResultSet getById(UUID id) {
         try (Connection connection = connectionDB()) {
             return connection
                     .createStatement()
-                    .executeQuery("SELECT * FROM users WHERE user_id = " + id);
+                    .executeQuery("SELECT * FROM users WHERE user_id = '" + id + "'");
         } catch (SQLException | IndexOutOfBoundsException e) {
             log.error("Ошибка при поиске человека по ID: " + e.getCause());
             return null;
@@ -56,8 +57,8 @@ public class UserRepository implements CrudRepository<User> {
             return connection
                     .createStatement()
                     .executeUpdate("INSERT INTO users (user_id, username, email, password, is_deleted, is_admin)" +
-                            " VALUES ('" + user.getUserId() + "', '" + user.getUserName() + "', '" + user.getEmail() + "', '"
-                            + user.getPassword() + "', '" + user.isDeleted() + "', '" + user.isAdmin() + "')");
+                            " VALUES ('" + user.getUserId() + "', '" + user.getUserName() + "', '" + user.getEmail() +
+                            "', '" + user.getPassword() + "', '" + user.isDeleted() + "', '" + user.isAdmin() + "')");
         } catch (SQLException e) {
             log.error("Ошибка при добавлении человека в СУБД: " + e.getCause());
             return 0;
@@ -70,10 +71,8 @@ public class UserRepository implements CrudRepository<User> {
             return connection
                     .createStatement()
                     .executeUpdate("UPDATE users SET " +
-                            "username = '" + user.getUserName() +
-                            "', email = '" + user.getEmail() +
-                            "', password = '" + user.getPassword() +
-                            "WHERE user_id = " + user.getUserId());
+                            "username = '" + user.getUserName() + "', email = '" + user.getEmail() +
+                            "', password = '" + user.getPassword() + "' WHERE user_id = '" + user.getUserId() + "'");
         } catch (SQLException e) {
             log.error("Ошибка при добавлении человека в СУБД: " + e.getCause());
             return 0;
@@ -81,14 +80,11 @@ public class UserRepository implements CrudRepository<User> {
     }
 
     @Override
-    public int delete(User user) {
+    public int delete(UUID id) {
         try (Connection connection = connectionDB()) {
             return connection
                     .createStatement()
-                    .executeUpdate("DELETE FROM users" +
-                            " WHERE username = '" + user.getUserName() +
-                            "' AND email = '" + user.getEmail() +
-                            "' AND password = '" + user.getPassword() + "'");
+                    .executeUpdate("DELETE FROM users WHERE user_id = '" + id + "'");
         } catch (SQLException e) {
             log.error("Ошибка при удалении человека с СУБД: " + e.getCause());
             return 0;
